+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 2.0.0 - 2023-04-06
+
+### Added
+
+* Redmine 5 Support (breaking change)
+* github templates
+* github actions to run automated tests
+
+### Removed
+
+* Chart.js from plugin in order to use the library incorporated in Redmine to version 3.7.1
+
 ## 1.1.1 - 2022-09-23
 
 ### Changed